+++ packages/ui/client/views/pluginsTemplate.js
@@ -8,7 +8,9 @@
 
 Template.pluginsTemplate.helpers({
 	templates: function() {
-		if(!Template[this.name]) return [];
+
+		if(!Template[this.name]) 
+			return [];
 
 		return K.Plugin.templatesByPlaceholder(this.name, Template.parentData());
 	}