+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
-- support custom style for each feature in K.Map.addGeojson()
-
+- support custom style for each geojson features in K.Map.addGeojson()
+- move features from plugin tracks to geoinfo under new method K.Geoinfo.getTrackInfo()
 
 ## v1.3.7
 - Meteor 1.7.0.5 compatibility