+++ CHANGELOG.md
@@ -9,7 +9,8 @@ Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 - K.Map.addGeojson new option clear
 - remove placeholder popupGeojson
 - added K.Place attribute classMarker
-- marker cursor is draggable
+- new option placehodelrs in K.Plugin definition
+- plugin categories support edits
 
 ## v1.3.7
 - Meteor 1.7.0.5 compatibility